SODA test customizing document key

+++ test/soda4.js
@@ -41,7 +41,7 @@ describe('168. soda4.js', () => {
     await sodaUtil.cleanup();
   });
 
-  it.skip('168.1 insertOneAndGet() method', async () => {
+  it('168.1 insertOneAndGet() fetches attributes without content', async () => {
     let conn;
     try {
       conn = await oracledb.getConnection(dbconfig);
@@ -65,8 +65,7 @@ describe('168. soda4.js', () => {
       (myDoc).should.be.an.Object();
 
       let content1 = myDoc.getContent();
-      console.log("getContent()");
-      console.log(content1);
+      should.not.exist(content1);
 
       // Fetch it back
       let doc2 = await coll.find().key(myKey).getOne();
@@ -92,14 +91,13 @@ describe('168. soda4.js', () => {
     }
   }); // 168.1
 
-  // ORA-40665:  A client-assigned key cannot be used for this operation.
-  it.skip('168.2 customize the key value', async () => {
-    let conn;
+  it('168.2 cannot customize the key value without metadata setting', async () => {
+    let conn, coll;
     try {
       conn = await oracledb.getConnection(dbconfig);
       let sd = conn.getSodaDatabase();
       let collectionName = 'soda_test_168_2';
-      let coll = await sd.createCollection(collectionName);
+      coll = await sd.createCollection(collectionName);
 
       // Insert a new document
       let testContent = {
@@ -107,23 +105,24 @@ describe('168. soda4.js', () => {
         address: {city: "Shenzhen", country: "China"}
       };
       let testKey = "86755";
-      let testDoc = await sd.createDocument(testContent, { key: testKey} );
+      let testDoc = sd.createDocument(testContent, { key: testKey} );
       should.exist(testDoc);
-      console.log(testDoc);
       await coll.insertOne(testDoc);
 
-      // Fetch it back
-      let doc2 = await coll.find().key(testKey).getOne();
-      let content2 = doc2.getContent();
-      console.log(content2);
-
-      await conn.commit();
-      let res = await coll.drop();
-      should.strictEqual(res.dropped, true);
-
     } catch(err) {
-      should.not.exist(err);
+      should.exist(err);
+      (err.message).should.startWith('ORA-40665:');
+      // ORA-40665: A client-assigned key cannot be used for this operation.
     } finally {
+      if (coll) {
+        try {
+          let res = await coll.drop();
+          should.strictEqual(res.dropped, true);
+        } catch(err) {
+          should.not.exist(err);
+        }
+      }
+
       if (conn) {
         try {
           await conn.close();
@@ -143,6 +142,7 @@ describe('168. soda4.js', () => {
       let coll = await sd.createCollection(collectionName);
 
       // Insert a new document
+      // Content is empty
       let testContent = {};
 
       let myDoc = await coll.insertOneAndGet(testContent);
@@ -172,5 +172,86 @@ describe('168. soda4.js', () => {
       }
     }
   }); // 168.3
+
+  it('168.4 customize the key value', async () => {
+    let conn;
+    try {
+      conn = await oracledb.getConnection(dbconfig);
+      let sd = conn.getSodaDatabase();
+      let collectionName = 'soda_test_168_4';
+      let testMetaData = {
+        "schemaName" : dbconfig.user.toUpperCase(),
+        "tableName" : collectionName,
+        "keyColumn" :
+                       {
+                         "name" : "ID",
+                         "sqlType" : "NUMBER",
+                         "assignmentMethod" : "CLIENT"
+                       },
+        "contentColumn" :
+                       {
+                         "name" : "JSON_DOCUMENTS",
+                         "sqlType" : "BLOB",
+                         "compress" : "NONE",
+                         "cache" : true,
+                         "encrypt" : "NONE",
+                         "validation" : "STRICT"
+                       },
+        "versionColumn" :
+                       {
+                         "name" : "VERSION",
+                         "type":"String",
+                         "method":"SHA256"
+                       },
+        "lastModifiedColumn" :
+                       {
+                         "name":"LAST_MODIFIED"
+                       },
+        "creationTimeColumn":
+                       {
+                         "name":"CREATED_ON"
+                       },
+        "readOnly": false
+      };
+
+      let coll = await sd.createCollection(collectionName, { metaData: testMetaData});
+
+      let testContent = {
+        name:    "Shelly",
+        address: {city: "Shenzhen", country: "China"}
+      };
+      
+      /* The key must always be a string and is always returned a string as 
+         well -- even if the "type" in the database is numeric. */
+      let testKey = '86755';
+      let testDoc = sd.createDocument(testContent, { key: testKey } );
+      should.strictEqual(testDoc.key, testKey);
+      await coll.insertOne(testDoc);
+
+      // Fetch it back
+      let docGot = await coll.find().key(testKey).getOne();
+      let contentGot = docGot.getContent();
+      should.strictEqual(contentGot.name, testContent.name);
+      should.strictEqual(
+        contentGot.address.country, 
+        testContent.address.country
+      );
+
+      await conn.commit();
+      let res = await coll.drop();
+      should.strictEqual(res.dropped, true);
+
+    } catch(err) {
+      should.not.exist(err);
+    } finally {
+      if (conn) {
+        try {
+          await conn.close();
+        } catch(err) {
+          should.not.exist(err);
+        }
+      }
+    }
+  }); // 168.4
 
 });
